Refactoring Avatar
