+++ superformula-frontend/src/components/Avatar/Avatar.tsx
@@ -1,9 +1,18 @@
 import React, { ReactElement } from 'react'
-import { AvatarProps } from './Avatar.types'
 import styles from './Avatar.module.scss'
 
-function Avatar({ imgUrl, imgAlt, className }: AvatarProps): ReactElement {
+type AvatarProps = {
+    imgUrl: string
+    imgAlt: string
+    className?: string
+}
+function Avatar({ imgUrl, imgAlt, className = '' }: AvatarProps): ReactElement {
     return <img className={`${styles.avatar} ${className}`} src={imgUrl} alt={imgAlt} />
 }
 
+Avatar.defaultProps = {
+    className: '',
+}
+
 export default Avatar
+export type { AvatarProps }
