--- superformula-frontend/src/components/Avatar/Avatar.types.tsx
@@ -1,7 +1,0 @@
-type AvatarProps = {
-    imgUrl: string
-    imgAlt: string
-    className?: string
-}
-
-export type { AvatarProps }
