+++ superformula-frontend/src/components/Avatar/index.tsx
@@ -1,0 +1,1 @@
+export * from './Avatar'
